# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Triplet API",
  "description": "Trip planning aplication API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/guest/lookup/{access_code}": {
   "get": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Lookup Guest Code",
    "description": "Which trip a link's code is for, so its page can ask for the PIN by name. The title only.",
    "operationId": "lookup_guest_code_guest_lookup__access_code__get",
    "parameters": [
     {
      "name": "access_code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Access Code"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestLookup"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/guest/access": {
   "post": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Guest Access",
    "operationId": "guest_access_guest_access_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GuestAccessCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestToken"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/guest/trip": {
   "get": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Get Guest Trip",
    "operationId": "get_guest_trip_guest_trip_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TripResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/guest/activities": {
   "get": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Get Guest Activities",
    "operationId": "get_guest_activities_guest_activities_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ActivityResponse"
         },
         "type": "array",
         "title": "Response Get Guest Activities Guest Activities Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Create Guest Activity",
    "operationId": "create_guest_activity_guest_activities_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActivityCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActivityResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/guest/itinerary": {
   "get": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Get Guest Itinerary",
    "operationId": "get_guest_itinerary_guest_itinerary_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestItineraryResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/guest/activities/{activity_id}": {
   "patch": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Update Guest Activity",
    "operationId": "update_guest_activity_guest_activities__activity_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "activity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Activity Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActivityUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActivityResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Delete Guest Activity",
    "operationId": "delete_guest_activity_guest_activities__activity_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "activity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Activity Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/guest-access": {
   "put": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Set Guest Access",
    "operationId": "set_guest_access_trips__trip_id__guest_access_put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GuestAccessSetup"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestAccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Update Guest Access",
    "description": "Show or hide costs, or let guests edit or not, without issuing a new code, so links already sent keep working.",
    "operationId": "update_guest_access_trips__trip_id__guest_access_patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GuestAccessUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestAccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Get Guest Access",
    "operationId": "get_guest_access_trips__trip_id__guest_access_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GuestAccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Guest Access"
    ],
    "summary": "Revoke Guest Access",
    "operationId": "revoke_guest_access_trips__trip_id__guest_access_delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ActivityCreate": {
    "properties": {
     "title": {
      "type": "string",
      "maxLength": 255,
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "location": {
      "type": "string",
      "maxLength": 255,
      "title": "Location"
     },
     "start_time": {
      "type": "string",
      "format": "date-time",
      "title": "Start Time"
     },
     "end_time": {
      "type": "string",
      "format": "date-time",
      "title": "End Time"
     },
     "estimated_cost": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 99999999.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Cost"
     },
     "source_link_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Link Id"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     }
    },
    "type": "object",
    "required": [
     "title",
     "location",
     "start_time",
     "end_time"
    ],
    "title": "ActivityCreate"
   },
   "ActivityResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "trip_id": {
      "type": "integer",
      "title": "Trip Id"
     },
     "source_link_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Link Id"
     },
     "place_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "location": {
      "type": "string",
      "title": "Location"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     },
     "start_time": {
      "type": "string",
      "format": "date-time",
      "title": "Start Time"
     },
     "end_time": {
      "type": "string",
      "format": "date-time",
      "title": "End Time"
     },
     "estimated_cost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Cost"
     }
    },
    "type": "object",
    "required": [
     "id",
     "trip_id",
     "title",
     "location",
     "start_time",
     "end_time"
    ],
    "title": "ActivityResponse"
   },
   "ActivityUpdate": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "location": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "start_time": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Time"
     },
     "end_time": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Time"
     },
     "estimated_cost": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 99999999.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Cost"
     },
     "source_link_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Link Id"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     }
    },
    "type": "object",
    "title": "ActivityUpdate"
   },
...
   "GuestAccessCreate": {
    "properties": {
     "access_code": {
      "type": "string",
      "maxLength": 16,
      "title": "Access Code"
     },
     "pin": {
      "type": "string",
      "maxLength": 12,
      "title": "Pin"
     }
    },
    "type": "object",
    "required": [
     "access_code",
     "pin"
    ],
    "title": "GuestAccessCreate"
   },
...
   "GuestAccessSetup": {
    "properties": {
     "pin": {
      "type": "string",
      "pattern": "^[A-Za-z0-9]{6,12}$",
      "title": "Pin"
     },
     "expires_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "show_costs": {
      "type": "boolean",
      "title": "Show Costs",
      "default": false
     },
     "allow_edits": {
      "type": "boolean",
      "title": "Allow Edits",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "pin"
    ],
    "title": "GuestAccessSetup"
   },
...
   "GuestItineraryResponse": {
    "properties": {
     "trip_id": {
      "type": "integer",
      "title": "Trip Id"
     },
     "days": {
      "items": {
       "$ref": "#/components/schemas/GuestItineraryDay"
      },
      "type": "array",
      "title": "Days"
     },
     "conflict_count": {
      "type": "integer",
      "title": "Conflict Count"
     },
     "allow_edits": {
      "type": "boolean",
      "title": "Allow Edits",
      "default": false
     },
     "show_costs": {
      "type": "boolean",
      "title": "Show Costs",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "trip_id",
     "days",
     "conflict_count"
    ],
    "title": "GuestItineraryResponse"
   },
   "GuestLookup": {
    "properties": {
     "title": {
      "type": "string",
      "title": "Title"
     },
     "trip_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trip Id"
     }
    },
    "type": "object",
    "required": [
     "title"
    ],
    "title": "GuestLookup",
    "description": "What a code's link page shows before the PIN is asked for: the title only."
   },
   "GuestToken": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type"
     },
     "refresh_token": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Refresh Token"
     },
     "access_code": {
      "type": "string",
      "title": "Access Code"
     },
     "show_costs": {
      "type": "boolean",
      "title": "Show Costs"
     },
     "allow_edits": {
      "type": "boolean",
      "title": "Allow Edits"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "token_type",
     "access_code",
     "show_costs",
     "allow_edits"
    ],
    "title": "GuestToken"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TripResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "destination": {
      "type": "string",
      "title": "Destination"
     },
     "destinations": {
      "items": {
       "$ref": "#/components/schemas/Destination"
      },
      "type": "array",
      "title": "Destinations",
      "default": []
     },
     "start_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "end_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Date"
     },
     "budget": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Budget"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "destination",
     "currency"
    ],
    "title": "TripResponse"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}